test(npm): move project-domain tests to test_project.py
Resolves https://github.com/hermetoproject/hermeto/issues/1156

Move TestPackage, TestPackageLock, and TestPurlifier from the mixed
npm test module into tests/unit/package_managers/npm/test_project.py.

This keeps npm tests aligned with module boundaries used by yarn and
yarn_classic.

No behavior changes; only test relocation and import adjustments.
